+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
-## [2.5.0] - TBD
+## [2.6.0] - TBD
+
+
+## [2.5.0] - 2023-03-10
+#### Feature
+- Insert BotConfig on next Leg
+- Support for Player Options
+
 #### Changed
 - Removed bots and busts from Global statistics
 
 #### Fixed
 - Correctly calculate Gotcha statistics
+- Ensure x01 statistics only count x01 legs
 
 ## [2.4.0] - 2022-10-21
 #### Feature
@@ -115,7 +123,8 @@
 #### Feature
 - Intial version of API for [kcapp-frontend](https://github.com/kcapp/frontend)
 
-[2.5.0]: https://github.com/kcapp/api/compare/v2.4.0...develop
+[2.6.0]: https://github.com/kcapp/api/compare/v2.5.0...develop
+[2.5.0]: https://github.com/kcapp/api/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/kcapp/api/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/kcapp/api/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/kcapp/api/compare/v2.1.0...v2.2.0